Merge 'bliss_msgs' package into 'bliss'
The custom interfaces generated are used exclusively by the `bliss`
package, so there was no real justification for `bliss_msgs` to be its
own package, other than for *good practices*. Moreover, when cleaning
the CMake cache, building both `bliss` and `bliss_msgs` at the same time
resulted in an error, because `bliss_msgs` couldn't be found (because it
didn't have the CMake config file to specify where to find it). Merging
the interfaces into `bliss` removes this issue.