Add 'time held' feature to D-Pad and buttons
This commit adds the time (ns) a button or D-Pad axis has been held for.
The value resets on release.

NOTES:

I observed a "peculiar" phenomenon with the D-Pad:

- Be buttons `n` from axis `N`, and `m` from axis `M != N`;
- Firstly, hold `n`;
- Secondly, hold `m`;
- Finally, release `n`;
- Result: `n`'s `raw`, and implicitly it's `time_held`, are NOT reset
  until either `m` is released, or the button opposite to `n` is pressed.

My hypothesis is that the cause of this phenomenon stems from the
hardware itself, and therefore WILL NOT, and CANNOT be fixed. This issue
is considered "known" and "unfixable". A warning WILL be added to the
documentation.